// web/src/api/models/search-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FulltextSearchParams } from './fulltext-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { GraphSearchParams } from './graph-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { SearchResultItem } from './search-result-item';
// May contain unused imports in some cases
// @ts-ignore
import type { SummarySearchParams } from './summary-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { VectorSearchParams } from './vector-search-params';
// May contain unused imports in some cases
// @ts-ignore
import type { VisionSearchParams } from './vision-search-params';

/**
 * 
 * @export
 * @interface SearchResult
 */
export interface SearchResult {
    /**
     * The id of the search result
     * @type {string}
     * @memberof SearchResult
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchResult
     */
    'query'?: string;
    /**
     * 
     * @type {VectorSearchParams}
     * @memberof SearchResult
     */
    'vector_search'?: VectorSearchParams;
    /**
     * 
     * @type {FulltextSearchParams}
     * @memberof SearchResult
     */
    'fulltext_search'?: FulltextSearchParams;
    /**
     * 
     * @type {GraphSearchParams}
     * @memberof SearchResult
     */
    'graph_search'?: GraphSearchParams;
    /**
     * 
     * @type {SummarySearchParams}
     * @memberof SearchResult
     */
    'summary_search'?: SummarySearchParams;
    /**
     * 
     * @type {VisionSearchParams}
     * @memberof SearchResult
     */
    'vision_search'?: VisionSearchParams;
    /**
     * 
     * @type {Array<SearchResultItem>}
     * @memberof SearchResult
     */
    'items'?: Array<SearchResultItem>;
    /**
     * The creation time of the search result
     * @type {string}
     * @memberof SearchResult
     */
    'created'?: string;
}